Implement coding of C++ namespaces.
To do this, the CppGenerator has to work more with objects and less with strings(which is a Good Thing(tm) anyway),
so add AdeRelationBase::GetPartner() which returns a pointer to a freshly
allocated AdeClass element.
Furthermore, add a functor AdeElementCompare() to sort any STL-container holding
pointers to generic model elements.

I'm not entirely happy with the optical appearance of the generated code, but I
consider it more important getting the code to work correctly...

TODO:
ResourceEdit must learn about namespaces in order to suggest reasonable
implementations of associations/aggregations/compositions; the CppGenerator
shouldn't mess with the contents of this field itself IMHO.


git-svn-id: http://astade.tigris.org/svn/astade/trunk@2171 aaa30fad-34f9-0310-8aa9-877cc762ee2e