Simplify AstadeFrame::InsertPrototype()
If the parameter list was changed and the body only contained the prototype
signature line and nothing else, the signature replacement logic would crash.
By simplifying said logic to only consider the very first line of the body,
the cause of the crash was fixed as well.

Test case:
- Add an operation, open and close code.cpp (but do not change anything).
- Add parameters and a parameter, open code.cpp again.
This was enough to make Astade crash.

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2186 aaa30fad-34f9-0310-8aa9-877cc762ee2e