AstadeTree: Fix deletion of classes and statecharts.
When a class (or a statechart) is deleted, (try to) remove it from the active
component first.  Otherwise a zombie entry in the component will cause an
annoying (but harmless) pop-up when clicking on the component node.

TODO:
We should probably try harder and remove the class/statechart from ALL
components of the model...

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2187 aaa30fad-34f9-0310-8aa9-877cc762ee2e