Fix deletion of self-relations.
When deleting a class which had a relation to itself (e.g. a node of a linked
list), Astade deletes the relation and the corresponding inrelation separately.
When deleting the inrelation, the relation will be deleted automatically, so
when the relation is deleted, the relation_*.ini file is already gone and the
partner name is thus unknown.

Deleting a nameless model element unfortunately acts as a recursive delete
starting at the model root, so the bug caused the entire model to be wiped out.

The nastiness of this bug will be hard to surpass...

Fixed by checking for the existence of the relation partner before attempting to
delete it.

Reported-by: Tim Reilly <reilly.t1@comcast.net>

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2190 aaa30fad-34f9-0310-8aa9-877cc762ee2e