Fixed the copying of relations.
If a class is copied from another model and it has a relation to a class which
also exists in the destination model, the corresponding inrelation would be
created _twice_ in the destination; one pointing to the (copied) relation
and one pointing to the source relation (in the _source_ model).

Deleting this second inrelation would wipe out the relation in the _source_
model, which would come as a surprise to most users...

Fixed by checking (in the AdeRelation constructor) that the partner lives in
the same model (and refusing to create the partner if not).

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2193 aaa30fad-34f9-0310-8aa9-877cc762ee2e