Set the time of modification to the generated files.
The coders now force the time of the generated files to be equal to the time
of the last modification of the class.

The header gets the last time of the model elements excluding action code;
the source gets the last time of all of the class sub-elements.

This should make the traffic light and exclamation mark behave as expected
across all operating systems as well as ease the stress on make, since a
header doesn't get a new time if only the action code was modified.

CAVEAT: Compiles, but otherwise completely untested!

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2203 aaa30fad-34f9-0310-8aa9-877cc762ee2e