Fix AstadeGenerate check for StateChartCoder
AstadeGenerate previously always checked that the generic state-chart coder
"StateChartCoder" was available in the directory given with the -S parameter.

As this makes the generation of the unit-tests more difficult (we want to
test the freshly built state-chart coder, not the installed one), change
the test to only look for the state-chart coder actually needed.

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2227 aaa30fad-34f9-0310-8aa9-877cc762ee2e