Fixed copy/paste of classes and operations.
It turned out that copy/paste (or drag'n'drop) didn't copy all features of
a class nor of its operations, so try to make copying more complete.
In particular the "CodingLanguage" of the class wasn't copied (but the legacy
"CCoded" was, so nobody noticed), nor was "Constraint", "Deprecated" or
"Tracelevel" of the operations.

Caveat: I might have missed more features...

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2313 aaa30fad-34f9-0310-8aa9-877cc762ee2e