Changed the "wait" function in ACF_Hal for Linux.
The "wait" function inside ACF_Hal was implemented using "sleep". This is sub optimal in a multy thread environment. Even when all statmachines run in only one thread, it should be possible to send messages from other threads.
In this case the "sleeping" state machine thread has to "wakeup".
I did an implementation with "sem_timedwait" which is a timeout on a semaphore. This can be interrupted by calling "sem_post" on the semaphore.
ACF_sendMessage uses this to wakeup the statemachine thread.

I did not yet extendet tests, but it seems to work in the few tests I did.

git-svn-id: http://astade.tigris.org/svn/astade/trunk@2317 aaa30fad-34f9-0310-8aa9-877cc762ee2e