Fix cooperation with Subversion 1.7 and above
Starting with Subversion 1.7 the meta info is consolidated into a single .svn
subdirectory residing in the top-level project directory, instead of being
spread out in .svn directories under each and every project subdir.

The old heuristics simply checked for the existence of .svn in the top-level
model directory and now fails if the Astade model is itself a subdirectory.

Now we first check if Git is installed, and if yes, ask Subversion if he
thinks we have a proper Subversion hierarchy.
Only if he does not, we activate Git.